refactor: move logseq client to folder
