+++ src/services/logseq/client.ts
@@ -1,5 +1,5 @@
 import type { ILSPluginUser } from "@logseq/libs/dist/LSPlugin.user.js";
-import type { LogseqServiceClient as LogseqServiceWrapper } from "./interfaces.js";
+import type { LogseqServiceClient as LogseqServiceWrapper } from "../interfaces.js";
 
 export const generateLogseqClient = (): LogseqServiceWrapper => {
   const logseq = window.logseq as ILSPluginUser;
